requirements.txt
gradio>=4.25.0
openai>=1.3.7
pillow
requests
beautifulsoup4


